Use named results instead of ARTIFACTS array
Makes the use of results much nicer as the specific named results can be
used instead of indexes within the ARTIFACTS array. We do loose the name
of the TaskRun from the context for the sake of simplicity, so that
another parameter is not introduced to carry the name over to the
scripts.