Add acceptance tests
This commits bootstraps the acceptance test framework.

A single scenario is added to ensure an artifact created from a file,
can be restored to its original state.

This is not currently tied to the CI. To run it:
```
cd acceptance && go test -v ./...
```

The implementation executes the create/use functionality via containers
since that is the interface to the underlying scripts. In theory, if the
implementation changes from bash to go, for example, there should be no
need to change the acceptance tests.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>