Use embedded Tekton Pipeline
The bundle containing the Pipeline reference uses out-of-date Task
bundles for the git-clone and the show-summary Tasks. They are old
enough that they are no longer considered to be trusted by the
Enterprise Contract.

There is likely a bug in how that Pipeline is generated. As a
workaround, this commit switches over to using an embedded Pipeline
definition so we can specify newer bundles for the Tasks mentioned
above.

Signed-off-by: Luiz Carvalho <lucarval@redhat.com>