allow the use of empty DEBUG env var
In the case where you are configuring a pipeline with many tasks,
it can be handy to be able to turn on DEBUG for all tasks by setting
a single param value.

```
    - name: trustedArtifactsDebug
      description: Flag to enable debug logging in trusted artifacts. Set to a non-empty string to enable.
      type: string
      default: ""
```

Task debugging can then be configured with this:

```
  stepTemplate:
    env:
      - name: DEBUG
        value: $(params.trustedArtifactsDebug)
```

Without this change, the mere setting of an empty DEBUG env var,
would turn on logging.

Signed-off-by: Scott Hebert <scoheb@gmail.com>