fix: retry attempts would always fail
Previously, scripts would call oras as below:

    retry.sh oras push --registry-config <(select-oci-auth.sh ${repo})

This uses process substitution to create an auth file on the fly from
the stdout of the select-oci-auth.sh command.

The process substitution will resolve to something like /dev/fd/63, so
the command ends up looking like

    retry.sh oras push --registry-config /dev/fd/63

But /dev/fd/63 is connected to the stdout of a process, so reading the
file once consumes the file.

Which means if the first 'oras push' fails for any reason, then all
subsequent attempts will fail with

    Error: failed to decode config file at /dev/fd/63:
    invalid config format: EOF

Fix this by using an actual file instead of a process substitution.

Signed-off-by: Adam Cmiel <acmiel@redhat.com>